fix: drop self-closing tag printing from the HTML printer
This drops the concept of "self-closing tags" and introduces logic for
dealing with void tags.

For example:

```html
// Before
<input type="text" />
// After (because input is a void element)
<input type="text">

// Before
<i class="fa" />
// After (because i isn't a valid void tag and shouldn't ever self-close)
<i class="fa"></i>
```

Currently, we self-close any tag without children. This is invalid and
has been for a very long time. Only void elements should be printed
without a closing tag, and all others should have one.